refactor(article-edit): replace diff tab with diff view component for improved content comparison
