--- apps/client/src/app/features/article/components/diff-tab/diff-tab.component.html
@@ -1,3 +1,0 @@
-<div class="placeholder">
-    <span class="placeholder-text">Сравнение изменений (в разработке)</span>
-</div>
