--- apps/client/src/app/features/article/components/diff-tab/diff-tab.component.scss
@@ -1,10 +1,0 @@
-.placeholder {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    flex: 1;
-}
-
-.placeholder-text {
-    color: var(--themed-text-secondary);
-}
