--- apps/client/src/app/features/article/components/diff-tab/diff-tab.component.ts
@@ -1,9 +1,0 @@
-import { ChangeDetectionStrategy, Component } from '@angular/core';
-
-@Component({
-    selector: 'app-diff-tab',
-    templateUrl: './diff-tab.component.html',
-    styleUrl: './diff-tab.component.scss',
-    changeDetection: ChangeDetectionStrategy.OnPush,
-})
-export class DiffTabComponent {}
