+++ apps/client/src/app/shared/components/cm-diff-view/cm-diff-view.component.scss
