+++ apps/client/src/app/shared/components/cm-diff-view/cm-diff-view.component.spec.ts
@@ -1,64 +1,24 @@
 import { LoggerService } from '@drevo-web/core';
 import { mockLoggerProvider, MockLoggerService } from '@drevo-web/core/testing';
-import { ApprovalStatus, VersionPairs } from '@drevo-web/shared';
 import { createComponentFactory, Spectator } from '@ngneat/spectator/jest';
-import { signal } from '@angular/core';
-import { DiffPageDataService } from '../../services/diff-page-data.service';
 import { CmDiffViewComponent } from './cm-diff-view.component';
-
-const mockVersionPairs: VersionPairs = {
-    current: {
-        articleId: 1,
-        versionId: 200,
-        content: 'new content',
-        author: 'Author A',
-        date: new Date('2025-01-15T14:30:00'),
-        title: 'Test Article',
-        info: 'Updated text',
-        approved: ApprovalStatus.Pending,
-        comment: '',
-    },
-    previous: {
-        articleId: 1,
-        versionId: 199,
-        content: 'old content',
-        author: 'Author B',
-        date: new Date('2025-01-14T10:00:00'),
-        title: 'Test Article',
-        info: '',
-        approved: ApprovalStatus.Approved,
-        comment: '',
-    },
-};
-
-function createMockDataService(
-    pairs: VersionPairs | undefined = mockVersionPairs,
-    error?: string
-): Partial<DiffPageDataService> {
-    return {
-        isLoading: signal(false).asReadonly(),
-        error: signal(error).asReadonly(),
-        versionPairs: signal(pairs).asReadonly(),
-    };
-}
 
 describe('CmDiffViewComponent', () => {
     let spectator: Spectator<CmDiffViewComponent>;
 
     const createComponent = createComponentFactory({
         component: CmDiffViewComponent,
-        providers: [
-            mockLoggerProvider(),
-            {
-                provide: DiffPageDataService,
-                useValue: createMockDataService(),
-            },
-        ],
+        providers: [mockLoggerProvider()],
         detectChanges: false,
     });
 
     beforeEach(() => {
-        spectator = createComponent();
+        spectator = createComponent({
+            props: {
+                oldText: 'old content',
+                newText: 'new content',
+            },
+        });
     });
 
     it('should create', () => {
