+++ apps/client/src/app/shared/components/diff-view/diff-view.component.scss
